+++ NEWS.md
@@ -7,6 +7,7 @@
 ## New features
 
 - Added new `summary.gtfs` method. Thanks Mark Padgham (@mapdge).
+- `import_gtfs()` has a new `encoding` parameter, used to handle encoded strings in their native encoding.
 
 ## Notes